requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.11.2
django-phonenumber-field==8.0.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.3.1
future==1.0.0
gunicorn==22.0.0
iso8601==2.1.0
mysqlclient==2.2.7
packaging==24.1
phonenumbers==9.0.2
pillow==11.2.1
PyJWT==2.8.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyMySQL==1.1.1
pyserial==3.5
python-decouple==3.8
pyusb==1.2.1
PyYAML==6.0.2
serial==0.0.97
sqlparse==0.5.3
typing_extensions==4.13.0
tzdata==2025.2
validate-email-address==1
whitenoise==6.9.0





